renaming to `breathing_earth_system_simulator` due to name conflict on PyPi
