--- BESS/version.txt
@@ -1,1 +1,0 @@
-1.0.1
