+++ breathing_earth_system_simulator/NDVI_minimum.jpeg
