+++ breathing_earth_system_simulator/SZA/UTC_to_solar.py
