+++ breathing_earth_system_simulator/__init__.py
@@ -1,5 +1,5 @@
 from .FLiESANN import *
-from .BESS import *
+from .breathing_earth_system_simulator import *
 
 from os.path import join, dirname, abspath
