+++ breathing_earth_system_simulator/canopy_shortwave_radiation.py
