+++ breathing_earth_system_simulator/carbon_uptake_efficiency.tif
